% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anolisMulti.R
\docType{data}
\name{anolisMulti}
\alias{anolisMulti}
\title{Anole data with a multiphylo object}
\format{An object of class \code{list} of length 2.}
\usage{
data(anolisMulti)
}
\description{
Anole data for aRbor functions
}
\keyword{datasets}
