% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeplyr_functions.R
\name{.as_data_frame.data.frame}
\alias{.as_data_frame.data.frame}
\title{as_data_frame.data.frame function from the tibble package}
\usage{
.as_data_frame.data.frame(x, ...)
}
\description{
as_data_frame.data.frame function from the tibble package
}
\keyword{internal}
