% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeplyr_functions.R
\name{.check_names_df}
\alias{.check_names_df}
\title{.check_names_df function from tibble package}
\usage{
.check_names_df(x, j)
}
\description{
.check_names_df function from tibble package
}
\keyword{internal}
