% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{.identifyBranches}
\alias{.identifyBranches}
\title{Internal function taken from bayou}
\usage{
.identifyBranches(tree, n, fixed.loc = TRUE, plot.simmap = TRUE)
}
\description{
Internal function taken from bayou
}
\keyword{internal}
