% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{.pars2map}
\alias{.pars2map}
\title{New version of .pars2map is faster, returns 3 elements rather than 2 named elements}
\usage{
.pars2map(pars, cache)
}
\description{
New version of .pars2map is faster, returns 3 elements rather than 2 named elements
}
\keyword{internal}
