% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{.plotRegimes}
\alias{.plotRegimes}
\title{Internal functions for painted_clades}
\usage{
.plotRegimes(tree, col = NULL, lwd = 1, pal = rainbow, ...)
}
\description{
Internal functions for painted_clades
}
\keyword{internal}
