% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{.prepare.branches}
\alias{.prepare.branches}
\title{This is an internal function modified from geiger's function .prepare.bm.univariate for use with OU models.}
\usage{
.prepare.branches(tree, ...)
}
\description{
This is an internal function modified from geiger's function .prepare.bm.univariate for use with OU models.
}
\keyword{internal}
