% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{.tipregime}
\alias{.tipregime}
\title{Internal function to determine tip regimes:}
\usage{
.tipregime(pars, tree)
}
\description{
Internal function to determine tip regimes:
}
\keyword{internal}
