% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeplyr_functions.R
\name{filter.treeplyrdata}
\alias{filter.treeplyrdata}
\alias{filter.grouped_treeplyrdata}
\title{Function for filtering rows from an object of class \code{treeplyrdata}}
\usage{
\method{filter}{treeplyrdata}(.data, ...)

\method{filter}{grouped_treeplyrdata}(.data, ...)
}
\arguments{
\item{.data}{An object of class \code{treeplyrdata}}

\item{...}{Additional arguments to filter by}
}
\value{
An object of class \code{treeplyrdata} with the dataset filtered by the specified criteria.
}
\description{
This function can be used to select a subset of species (rows) from a treeplyrdata object;
see \code{\link{filter}}.
}
\examples{
data(anolis)
td <- make.treeplyrdata(anolis$phy, anolis$dat, name_column=1)
tdfilter <- filter(td, island=="Cuba", SVL > 3.5)
}
\seealso{
\code{\link{filter}}
}
