% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeplyr_functions.R
\name{filter_-defunct}
\alias{filter_-defunct}
\alias{filter_.treeplyrdata}
\alias{filter_.grouped_treeplyrdata}
\title{filter_.treeplyrdata}
\usage{
\method{filter_}{treeplyrdata}(.data, ...)

\method{filter_}{grouped_treeplyrdata}(.data, ...)
}
\description{
Now defunct and replaced with filter.treeplyrdata

Now defunct and replaced with filter.groupedtreeplyrdata.
}
\seealso{
\code{\link{treeplyr-defunct}}

\code{\link{treeplyr-defunct}}
}
\keyword{internal}
