% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectCharacterType.R
\name{filterMatrix}
\alias{filterMatrix}
\title{Filter a matrix, returning either all continuous or all discrete characters}
\usage{
filterMatrix(mat, charType, returnType = "discrete")
}
\arguments{
\item{mat}{A matrix of data}

\item{charType}{A vector of character types (perhaps from detectAllCharacters)}

\item{returnType}{Either discrete or continuous}
}
\value{
Matrix with only discrete or continuous characters
}
\description{
Filter a matrix, returning either all continuous or all discrete characters
}
\examples{
data(anolis)
aType<-detectAllCharacters(anolis$dat)
filterMatrix(anolis$dat, aType, "discrete")
}
