% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectCharacterType.R
\name{forceNames}
\alias{forceNames}
\title{Force names for rows, columns, or both}
\usage{
forceNames(dat, nameType = "row")
}
\arguments{
\item{dat}{A vector of data}

\item{nameType, }{either:
\describe{
    \item{"row"}{Rows}
         \item{"col"}{Columns}
         \item{"rowcol"}{Both rows and columns}
}}
}
\description{
Force names for rows, columns, or both
}
\examples{
data(anolis)
forceNames(anolis$dat, "row")
}
