% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeplyr_functions.R
\name{getVector}
\alias{getVector}
\title{A function for returning a named vector from a data frame or matrix with row names}
\usage{
getVector(td, ...)
}
\arguments{
\item{td}{A treedata object}

\item{...}{The name of the column to select}
}
\value{
A named vector
}
\description{
A function for returning a named vector from a data frame or matrix with row names
}
