% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeplyr_functions.R
\name{group_by_-defunct}
\alias{group_by_-defunct}
\title{group_by_.treedata}
\description{
Now defunct and replaced with group_by.treedata.
}
\seealso{
\code{\link{treeplyr-defunct}}
}
\keyword{internal}
