% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectCharacterType.R
\name{hasNames}
\alias{hasNames}
\title{Row and column name check}
\usage{
hasNames(dat, nameType = "row")
}
\arguments{
\item{dat}{A vector of data}

\item{nameType, }{either:
\describe{
    \item{"row"}{Rows}
       \item{"col"}{Columns}
         \item{"rowcol"}{Both rows and columns}
}}
}
\description{
Row and column name check
}
\examples{
data(anolis)
hasNames(anolis$dat, "row")
}
