% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeplyr_functions.R
\name{make.treedata}
\alias{make.treedata}
\title{Function for making an object of class \code{treedata}}
\usage{
make.treedata(phy, data, name_column = "detect")
}
\arguments{
\item{phy}{An object of class 'phylo' or 'multiPhylo}

\item{data}{A data frame or matrix}

\item{name_column}{An optional argument that specifies the column of \code{data} that contains the 
names to be matched to the tree. By default, it is set to "detect" which finds the column with the 
most matches to the tree (including 
the rownames).}
}
\value{
An object of class "\code{treedata}". The tree is pruned of tips not represented in the data, 
and the data is filtered for taxa not in the tree. The data is returned as a data frame tble that is 
compatible with \code{dplyr} functions.
}
\description{
This function generates an object of class \code{treedata} that ensures that the ordering of tip labels 
and data remain intact. The object can be manipulated using \code{dplyr} functions.
}
\examples{
data(anolis)
td <- make.treedata(anolis$phy, anolis$dat)

# Using multiPhylo objects

data(anolis)
a1<-anolis$phy
a1$tip.label[1]<-"NA"
trees<-list(a1, anolis$phy)
class(trees)<-"multiPhylo"
td<-make.treedata(phy=trees, data=anolis$dat[-c(5:10),])
}
