% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeplyr_functions.R
\name{mutate.treeplyrdata}
\alias{mutate.treeplyrdata}
\alias{mutate.grouped_treeplyrdata}
\title{Function for mutating an object of class \code{treeplyrdata}}
\usage{
\method{mutate}{treeplyrdata}(.data, ...)

\method{mutate}{grouped_treeplyrdata}(.data, ...)
}
\arguments{
\item{.data}{An object of class \code{treeplyrdata}}

\item{...}{Arguments to mutate the treeplyrdata object}
}
\value{
An object of class \code{treeplyrdata} with new data added.
}
\description{
This function can be used to add new variables to a treeplyrdata object; see \code{\link{mutate}}.
}
\examples{
data(anolis)
td <- make.treeplyrdata(anolis$phy, anolis$dat)
tdmutate <- mutate(td, lnSVL = log(SVL), badassery = awesomeness + hostility)
}
\seealso{
\code{\link{mutate}}
}
