% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeplyr_functions.R
\name{mutate_-defunct}
\alias{mutate_-defunct}
\alias{mutate_.treedata}
\title{mutate_.treedata}
\usage{
\method{mutate_}{treedata}(.data, ...)
}
\description{
Now defunct and replaced with mutate.treedata and mutate_.grouped_treedata.
}
\seealso{
\code{\link{treeplyr-defunct}}
}
\keyword{internal}
