% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeplyr_functions.R
\name{paint_clades}
\alias{paint_clades}
\title{Add regimes to a treedata object}
\usage{
paint_clades(tdObject, nclades, name = "clades", interactive = TRUE,
  type = "nodes", ids = NULL, ...)
}
\description{
Add regimes to a treedata object
}

