% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeplyr_functions.R
\name{s_filter}
\alias{s_filter}
\title{Modified version of dplyr's filter that uses string arguments}
\usage{
s_filter(.data, ...)
}
\arguments{
\item{.data}{dplyr object}

\item{...}{Additional arguments}
}
\description{
Modified version of dplyr's filter that uses string arguments
}

