% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeplyr_functions.R
\name{s_summarise}
\alias{s_summarise}
\title{Modified version of dplyr's summarise that uses string arguments}
\usage{
s_summarise(.data, ...)
}
\arguments{
\item{.data}{dplyr object}

\item{...}{Additional arguments}
}
\description{
Modified version of dplyr's summarise that uses string arguments
}

