% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeplyr_functions.R
\name{s_treeply}
\alias{s_treeply}
\title{Modified version of dplyr's group_by that uses string arguments}
\usage{
s_treeply(.data, ...)
}
\arguments{
\item{.data}{dplyr object}

\item{...}{Additional arguments}
}
\description{
Modified version of dplyr's group_by that uses string arguments
}

