% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeplyr_functions.R
\name{select.treeplyrdata}
\alias{select.treeplyrdata}
\alias{select.grouped_treeplyrdata}
\title{Function for selecting columns from an object of class \code{treeplyrdata}}
\usage{
\method{select}{treeplyrdata}(.data, ...)
}
\arguments{
\item{.data}{An object of class \code{treeplyrdata}}

\item{...}{Additional arguments to select columns}
}
\value{
An object of class \code{treeplyrdata} with specified variables selected.
}
\description{
This function can be used to select a subset of variables (columns) from a treeplyrdata object;
see \code{\link{select}}.
}
\examples{
data(anolis)
td <- make.treeplyrdata(anolis$phy, anolis$dat)
tdselect <- select(td, SVL, awesomeness)
}
\seealso{
\code{\link{select}}
}
