% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeplyr_functions.R
\name{slice.treeplyrdata}
\alias{slice.treeplyrdata}
\title{Choose rows by their ordinal position in the tbl for an object of class \code{treeplyrdata}}
\usage{
\method{slice}{treeplyrdata}(.data, ...)
}
\arguments{
\item{.data}{An object of class \code{treeplyrdata}}

\item{...}{Integer row values}
}
\value{
An object of class \code{treeplyrdata}.
}
\description{
This function can be used to drop tips from tree and data; see \code{\link{slice}}.
}
\examples{
data(anolis)
td <- make.treeplyrdata(anolis$phy, anolis$dat)
tdslice <- slice(td, 1:5)
tdslice
}
\seealso{
\code{\link{slice}}
}
