% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeplyr_functions.R
\name{summarise.treeplyrdata}
\alias{summarise.treeplyrdata}
\alias{summarize.treeplyrdata}
\alias{summarize.grouped_treeplyrdata}
\alias{summarise.grouped_treeplyrdata}
\title{Function for summarizing an object of class \code{treeplyrdata}}
\usage{
\method{summarise}{treeplyrdata}(.data, ...)

\method{summarise}{grouped_treeplyrdata}(.data, ...)
}
\arguments{
\item{.data}{An object of class \code{treeplyrdata}}

\item{...}{Additional expressions by which to summarize data in the \code{treeplyrdata} object}
}
\value{
An object of class \code{tbl_df} with the requested summary data.
}
\description{
This function can be used to summarize a treeplyrdata object.
}
\details{
Summarizing \code{treeplyrdata} objects allows expressions using the objects \code{phy}. The \code{treeplyrdata}
object can also be grouped, with summary statistics being applied to the pruned groups and phylogenies.
}
\examples{
data(anolis)
td <- make.treeplyrdata(anolis$phy, anolis$dat)
summarize(td, ntips = length(phy$tip.label), meanSVL = mean(SVL), sdSVL = sd(SVL))
tdGrouped <- group_by(td, ecomorph)
summarize(tdGrouped, ntips = length(phy$tip.label),
   totalBL = sum(phy$edge.length), meanSVL = mean(SVL), sdSVL = sd(SVL))
}
\seealso{
\code{\link{summarize}}, \code{\link{group_by}}
}
