% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeplyr_defunct.R
\name{treeplyr-defunct}
\alias{treeplyr-defunct}
\alias{group_by_}
\alias{group_by_.treedata}
\alias{mutate_}
\alias{mutate_.treedata}
\title{Defunct functions in treeplyr}
\description{
These functions have been removed to reflect changes in dplyr.
}
\details{
\itemize{
 \item \code{\link{group_by_.treedata}}: This function is defunct, use group_by instead.
 \item \code{\link{mutate_.treedata}}: This function is defunct, use mutate instead.
}
}
