% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeplyr.R
\docType{package}
\name{treeplyr}
\alias{treeplyr}
\alias{treeplyr-package}
\title{treeplyr: An R package that allows dplyr to be used on matched tree and data objects}
\description{
An R package that matches phylogenetic trees and trait data, and allows simultaneous dplyr-like 
manipulation of matched tree and/or data.
}
\author{
Josef Uyeda
}

