% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeplyr_functions.R
\name{filter_-defunct}
\alias{filter_-defunct}
\alias{filter_.treedata}
\alias{filter_.grouped_treedata}
\title{filter_.treedata}
\usage{
\method{filter_}{treedata}(.data, ...)

\method{filter_}{grouped_treedata}(.data, ...)
}
\description{
Now defunct and replaced with filter.treedata

Now defunct and replaced with filter.groupedtreedata.
}
\seealso{
\code{\link{treeplyr-defunct}}

\code{\link{treeplyr-defunct}}
}
\keyword{internal}
