% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeplyr_functions.R
\name{mutate_-defunct}
\alias{mutate_-defunct}
\alias{mutate_.treeplyrdata}
\alias{mutate_.grouped_treeplyrdata}
\title{mutate_.treeplyrdata}
\usage{
\method{mutate_}{treeplyrdata}(.data, ...)

\method{mutate_}{grouped_treeplyrdata}(.data, ..., add = FALSE)
}
\description{
Now defunct and replaced with mutate.treeplyrdata and mutate_.grouped_treeplyrdata.

Now defunct and replaced with mutate.groupedtreeplyrdata.
}
\seealso{
\code{\link{treeplyr-defunct}}

\code{\link{treeplyr-defunct}}
}
\keyword{internal}
