% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeplyr_functions.R
\name{select.treedata}
\alias{select.treedata}
\alias{select.grouped_treedata}
\title{Function for selecting columns from an object of class \code{treedata}}
\usage{
\method{select}{treedata}(.data, ...)
}
\arguments{
\item{.data}{An object of class \code{treedata}}

\item{...}{Additional arguments to select columns}
}
\value{
An object of class \code{treedata} with specified variables selected.
}
\description{
This function can be used to select a subset of variables (columns) from a treedata object;
see \code{\link{select}}.
}
\examples{
data(anolis)
td <- make.treedata(anolis$phy, anolis$dat)
tdselect <- select(td, SVL, awesomeness)
}
\seealso{
\code{\link{select}}
}
