% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeplyr_functions.R
\name{select_-defunct}
\alias{select_-defunct}
\alias{select_.treedata}
\title{select_.treedata}
\usage{
\method{select_}{treedata}(.data, ...)
}
\description{
Now defunct and replaced with select.treedata
}
\seealso{
\code{\link{treeplyr-defunct}}
}
\keyword{internal}
