% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeplyr_functions.R
\name{slice.treedata}
\alias{slice.treedata}
\title{Choose rows by their ordinal position in the tbl for an object of class \code{treedata}}
\usage{
\method{slice}{treedata}(.data, ...)
}
\arguments{
\item{.data}{An object of class \code{treedata}}

\item{...}{Integer row values}
}
\value{
An object of class \code{treedata}.
}
\description{
This function can be used to drop tips from tree and data; see \code{\link{slice}}.
}
\examples{
data(anolis)
td <- make.treedata(anolis$phy, anolis$dat)
tdslice <- slice(td, 1:5)

# Using multiPhylo objects

data(anolis)
a1<-anolis$phy
a1$tip.label[1]<-"NA"
trees<-list(a1, anolis$phy)
class(trees)<-"multiPhylo"
td<-make.treedata(phy=trees, data=anolis$dat[-c(5:10),])
tdslice <- slice(td, 1:5)
}
\seealso{
\code{\link{slice}}
}
