% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeplyr_functions.R
\name{slice_-defunct}
\alias{slice_-defunct}
\alias{slice_.treedata}
\title{slice_.treedata}
\usage{
\method{slice_}{treedata}(.data, ...)
}
\description{
Now defunct and replaced with slice.treedata
}
\seealso{
\code{\link{treeplyr-defunct}}
}
\keyword{internal}
