% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeplyr_defunct.R
\name{treeplyr-defunct}
\alias{treeplyr-defunct}
\title{Defunct functions in treeplyr}
\description{
These functions have been removed to reflect changes in dplyr.
}
\details{
\itemize{
 \item \code{\link{group_by_.treedata}}: This function is defunct, use group_by instead.
 \item \code{\link{mutate_.treedata}}, \code{\link{mutate_.grouped_treedata}}: This function is defunct, use mutate instead.
 \item \code{\link{slice_.treedata}}: This function is defunct, use slice instead.
 \item \code{\link{select_.treedata}}: This function is defunct, use select instead.
 \item \code{\link{filter_.treedata}},\code{\link{filter_.grouped_treedata}}: This function is defunct, use filter instead.
}
}
