% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeplyr_defunct.R, R/treeplyr_functions.R
\name{treeplyr-defunct}
\alias{treeplyr-defunct}
\alias{group_by_}
\alias{group_by_.treedata}
\alias{mutate_}
\alias{mutate_.treedata}
\title{Defunct functions in treeplyr}
\usage{
\method{mutate_}{treedata}(.data, ...)

\method{group_by_}{treedata}(.data, ..., .dots = list(), add = FALSE)
}
\arguments{
\item{.data}{A data frame}

\item{.dots, ...}{Pair/valyes of expressions coercible to lazy objects}

\item{add}{Add to or override existing groups}
}
\description{
These functions have been removed to reflect changes in dplyr.
}
\details{
\itemize{
 \item \code{\link{group_by_.treedata}}: This function is defunct, use group_by instead.
 \item \code{\link{mutate_.treedata}}: This function is defunct, use mutate instead.
}
}
\section{\code{mutate_.treedata}}{

Instead of \code{mutate_}, use \code{\link{mutate}}.
}

\section{\code{group_by_.treedata}}{

Instead of \code{group_by_}, use \code{\link{group_by}}.
}

