% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeplyr_defunct.R
\name{treeplyr-defunct}
\alias{treeplyr-defunct}
\title{Defunct functions in treeplyr}
\description{
These functions have been removed to reflect changes in dplyr.
}
\details{
\itemize{
 \item \code{\link{group_by_.treedata}}: This function is defunct.
}
}
