% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeplyr-package.R
\docType{package}
\name{treeplyr}
\alias{treeplyr}
\title{treeplyr: 'dplyr' Functionality for Matched Tree and Data Objects}
\description{
Matches phylogenetic trees and trait data, and allows simultaneous 
manipulation of the tree and data using 'dplyr'.
}
\author{
Josef Uyeda
}
